backend/requirements.txt
# MassFlow Backend - API
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic[email-validator]==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
python-dotenv==1.0.1
